Move connection handling to universe.js
