+++ src/components/editor-mode/scene-menu/SceneMenu.jsx
@@ -1,7 +1,6 @@
 import React, {useState, useEffect, useContext, useRef} from 'react';
 import classnames from 'classnames';
 
-import {world} from '../../../../world';
 import universe from '../../../../universe';
 import voiceInput from '../../../../voice-input/voice-input';
 import sceneNames from '../../../../scenes/scenes.json';
@@ -111,12 +110,12 @@ export const SceneMenu = ({
   const handleRoomSelect = room => {
     setState({openedPanel: null});
 
-    if (!world.isConnected()) {
+    if (!universe.isConnected()) {
       universe.pushUrl(
         `/?src=${encodeURIComponent(selectedScene)}&room=${room.name}`,
       );
 
-      /* const isConnected = world.isConnected();
+      /* const isConnected = universe.isConnected();
             setMultiplayerConnected(isConnected);
             if (isConnected) {
               setRoomName(room.name);
