+++ universe.js
@@ -4,18 +4,25 @@ responsibilities include loading the world on url change.
 */
 
 // import * as THREE from 'three';
-import * as Z from 'zjs';
-import {world} from './world.js';
-import physicsManager from './physics-manager.js';
-import {loadOverworld} from './overworld.js';
-import {initialPosY} from './constants.js';
-import {parseQuery} from './util.js';
 import metaversefile from 'metaversefile';
+import WSRTC from 'wsrtc/wsrtc.js';
+import * as Z from 'zjs';
+
+import {appsMapName, initialPosY, playersMapName} from './constants.js';
+import {loadOverworld} from './overworld.js';
+import physicsManager from './physics-manager.js';
+import physxWorkerManager from './physx-worker-manager.js';
+import physx from './physx.js';
+import {playersManager} from './players-manager.js';
+import {getLocalPlayer} from './players.js';
 import sceneNames from './scenes/scenes.json';
+import {parseQuery} from './util.js';
+import {world} from './world.js';
 
 class Universe extends EventTarget {
   constructor() {
     super();
+    this.wsrtc = null;
 
     this.currentWorld = null;
     this.sceneLoadedPromise = null;
@@ -25,7 +32,7 @@ class Universe extends EventTarget {
       ((window.location.port ? parseInt(window.location.port, 10) : (window.location.protocol === 'https:' ? 443 : 80)) + 1) + '/worlds/';
   }
   async enterWorld(worldSpec) {
-    world.disconnectRoom();
+    this.disconnectRoom();
     
     const localPlayer = metaversefile.useLocalPlayer();
     /* localPlayer.teleportTo(new THREE.Vector3(0, 1.5, 0), camera.quaternion, {
@@ -45,7 +52,7 @@ class Universe extends EventTarget {
       const {src, room} = worldSpec;
       if (!room) {
         const state = new Z.Doc();
-        world.connectState(state);
+        this.connectState(state);
         
         let match;
         if (src === undefined) {
@@ -68,7 +75,7 @@ class Universe extends EventTarget {
       } else {
         const p = (async () => {
           const roomUrl = this.getWorldsHost() + room;
-          await world.connectRoom(roomUrl);
+          await this.connectRoom(roomUrl);
         })();
         promises.push(p);
       }
@@ -117,6 +124,83 @@ class Universe extends EventTarget {
         });
       }
     }
+  }
+
+  isConnected() { return !!this.wsrtc; }
+
+  getConnection() { return this.wsrtc; }
+
+  // called by enterWorld() in universe.js
+  // This is called in single player mode instead of connectRoom
+  connectState(state) {
+    state.setResolvePriority(1);
+    playersManager.bindState(state.getArray(playersMapName));
+
+    world.appManager.unbindState();
+    world.appManager.clear();
+    const appsArray = state.get(appsMapName, Z.Array);
+
+    world.appManager.bindState(appsArray);
+
+    const localPlayer = getLocalPlayer();
+    localPlayer.bindState(state.getArray(playersMapName));
+  }
+
+  // called by enterWorld() in universe.js
+  // This is called when a user joins a multiplayer room
+  // either from single player or directly from a link
+  async connectRoom(u, state = new Z.Doc()) {
+    // Players cannot be initialized until the physx worker is loaded
+    // Otherwise you will receive allocation errors because the module instance is undefined
+    await physx.waitForLoad();
+    await physxWorkerManager.waitForLoad();
+    const localPlayer = getLocalPlayer();
+
+    state.setResolvePriority(1);
+
+    // Create a new instance of the websocket rtc client
+    // This comes from webaverse/wsrtc/wsrtc.js
+    this.wsrtc = new WSRTC(u, {
+      localPlayer,
+      crdtState: state,
+    });
+
+    // This is called when the websocket connection opens, i.e. server is connectedw
+    const open = e => {
+      this.wsrtc.removeEventListener('open', open);
+      // Clear the last world state
+      const appsArray = state.get(appsMapName, Z.Array);
+      playersManager.bindState(state.getArray(playersMapName));
+
+      // Unbind the world state to clear existing apps
+      world.appManager.unbindState();
+      world.appManager.clear();
+      // Bind the new state
+      world.appManager.bindState(appsArray);
+
+      // Called by WSRTC when the connection is initialized
+      const init = e => {
+        this.wsrtc.removeEventListener('init', init);
+        localPlayer.bindState(state.getArray(playersMapName));
+
+        this.wsrtc.addEventListener('audio', e => {
+          const player = playersManager.remotePlayersByInteger.get(e.data.playerId);
+          player.processAudioData(e.data);
+        });
+      };
+
+      this.wsrtc.addEventListener('init', init);
+    };
+
+    this.wsrtc.addEventListener('open', open);
+
+    return this.wsrtc;
+  }
+
+  // called by enterWorld() in universe.js, to make sure we aren't already connected
+  disconnectRoom() {
+    if (this.wsrtc && this.wsrtc.state === 'open') this.wsrtc.close();
+    this.wsrtc = null;
   }
 }
 const universe = new Universe();
